refactor: 키워드 제거, 사용하지 않는 dto 정리
