+++ backend/src/main/java/reviewme/review/controller/ReviewApi.java
@@ -16,7 +16,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
-import reviewme.review.dto.request.create.CreateReviewRequest;
+import reviewme.review.dto.request.CreateReviewRequest;
 import reviewme.review.dto.response.ReceivedReviewsResponse2;
 import reviewme.review.service.dto.response.detail.TemplateAnswerResponse;
