+++ backend/src/main/java/reviewme/review/controller/ReviewController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import reviewme.global.HeaderProperty;
-import reviewme.review.dto.request.create.CreateReviewRequest;
+import reviewme.review.dto.request.CreateReviewRequest;
 import reviewme.review.dto.response.ReceivedReviewsResponse2;
 import reviewme.review.service.CreateReviewService;
 import reviewme.review.service.ReviewDetailLookupService;
