+++ backend/src/main/java/reviewme/review/dto/request/CreateReviewAnswerRequest.java
@@ -1,4 +1,4 @@
-package reviewme.review.dto.request.create;
+package reviewme.review.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
