--- backend/src/main/java/reviewme/review/dto/response/ReviewContentResponse.java
@@ -1,17 +1,0 @@
-package reviewme.review.dto.response;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-
-@Schema(name = "리뷰 내용 응답")
-public record ReviewContentResponse(
-
-        @Schema(description = "리뷰 내용 ID")
-        long id,
-
-        @Schema(description = "리뷰 문항")
-        String question,
-
-        @Schema(description = "리뷰 문항에 대한 답변")
-        String answer
-) {
-}
