+++ backend/src/main/java/reviewme/review/service/CreateReviewService.java
@@ -13,8 +13,8 @@ import reviewme.review.domain.CheckboxAnswer;
 import reviewme.review.domain.Review2;
 import reviewme.review.domain.TextAnswer;
 import reviewme.review.domain.exception.ReviewGroupNotFoundByRequestReviewCodeException;
-import reviewme.review.dto.request.create.CreateReviewAnswerRequest;
-import reviewme.review.dto.request.create.CreateReviewRequest;
+import reviewme.review.dto.request.CreateReviewAnswerRequest;
+import reviewme.review.dto.request.CreateReviewRequest;
 import reviewme.review.repository.QuestionRepository2;
 import reviewme.review.repository.Review2Repository;
 import reviewme.review.service.exception.SubmittedQuestionAndProvidedQuestionMismatchException;
