+++ backend/src/test/java/reviewme/review/service/CreateReviewServiceTest.java
@@ -13,8 +13,8 @@ import reviewme.question.domain.Question2;
 import reviewme.question.domain.QuestionType;
 import reviewme.question.repository.OptionGroupRepository;
 import reviewme.question.repository.OptionItemRepository;
-import reviewme.review.dto.request.create.CreateReviewAnswerRequest;
-import reviewme.review.dto.request.create.CreateReviewRequest;
+import reviewme.review.dto.request.CreateReviewAnswerRequest;
+import reviewme.review.dto.request.CreateReviewRequest;
 import reviewme.review.repository.CheckboxAnswerRepository;
 import reviewme.review.repository.QuestionRepository2;
 import reviewme.review.repository.Review2Repository;
